+++ web/handler/session_test.go
@@ -22,484 +22,6 @@ import (
 	"github.com/labstack/echo/v4"
 )
 
-func TestSessionHandler_State(t *testing.T) {
-	tests := []struct {
-		name                     string
-		sessionID                string
-		body                     string
-		userID                   string
-		prepareMockPlayerFn      func(m *mock_spotify.MockPlayer)
-		prepareMockPusherFn      func(m *mock_event.MockPusher)
-		prepareMockUserRepoFn    func(m *mock_repository.MockUser)
-		prepareMockSessionRepoFn func(m *mock_repository.MockSession)
-		wantErr                  bool
-		wantCode                 int
-	}{
-		{
-			name:                     "存在しないstateのとき400",
-			sessionID:                "sessionID",
-			body:                     `{"state": "INVALID"}`,
-			prepareMockPlayerFn:      func(m *mock_spotify.MockPlayer) {},
-			prepareMockPusherFn:      func(m *mock_event.MockPusher) {},
-			prepareMockUserRepoFn:    func(m *mock_repository.MockUser) {},
-			prepareMockSessionRepoFn: func(m *mock_repository.MockSession) {},
-			wantErr:                  true,
-			wantCode:                 http.StatusBadRequest,
-		},
-		{
-			name:                  "PLAYで指定されたidのセッションが存在しないとき404",
-			sessionID:             "notFoundSessionID",
-			body:                  `{"state": "PLAY"}`,
-			prepareMockPlayerFn:   func(m *mock_spotify.MockPlayer) {},
-			prepareMockPusherFn:   func(m *mock_event.MockPusher) {},
-			prepareMockUserRepoFn: func(m *mock_repository.MockUser) {},
-			prepareMockSessionRepoFn: func(m *mock_repository.MockSession) {
-				m.EXPECT().FindByID("notFoundSessionID").Return(nil, entity.ErrSessionNotFound)
-			},
-			wantErr:  true,
-			wantCode: http.StatusNotFound,
-		},
-		{
-			// https://github.com/camphor-/relaym-client/issues/195
-			name:      "PLAYで再生するデバイスがオフラインかつStateがPauseかつセッション作成者のリクエストのときは403を返しつつも再生処理を行う",
-			sessionID: "sessionID",
-			body:      `{"state": "PLAY"}`,
-			userID:    "creator_id",
-			prepareMockPlayerFn: func(m *mock_spotify.MockPlayer) {
-				m.EXPECT().SetRepeatMode(gomock.Any(), false, "device_id").Return(entity.ErrActiveDeviceNotFound)
-				m.EXPECT().SetShuffleMode(gomock.Any(), false, "device_id").Return(entity.ErrActiveDeviceNotFound)
-				m.EXPECT().Play(gomock.Any(), "device_id").Return(nil)
-			},
-			prepareMockPusherFn: func(m *mock_event.MockPusher) {
-			},
-			prepareMockUserRepoFn: func(m *mock_repository.MockUser) {},
-			prepareMockSessionRepoFn: func(m *mock_repository.MockSession) {
-				m.EXPECT().FindByID("sessionID").Return(&entity.Session{
-					ID:        "sessionID",
-					Name:      "session_name",
-					CreatorID: "creator_id",
-					QueueHead: 0,
-					DeviceID:  "device_id",
-					StateType: entity.Pause,
-					QueueTracks: []*entity.QueueTrack{
-						{Index: 0, URI: "spotify:track:5uQ0vKy2973Y9IUCd1wMEF"},
-						{Index: 1, URI: "spotify:track:49BRCNV7E94s7Q2FUhhT3w"},
-					},
-				}, nil)
-				m.EXPECT().Update(&entity.Session{
-					ID:        "sessionID",
-					Name:      "session_name",
-					CreatorID: "creator_id",
-					QueueHead: 0,
-					DeviceID:  "device_id",
-					StateType: "PLAY",
-					QueueTracks: []*entity.QueueTrack{
-						{Index: 0, URI: "spotify:track:5uQ0vKy2973Y9IUCd1wMEF"},
-						{Index: 1, URI: "spotify:track:49BRCNV7E94s7Q2FUhhT3w"},
-					},
-				}).Return(nil)
-			},
-			wantErr:  true,
-			wantCode: http.StatusForbidden,
-		},
-		{
-			name:                  "不正なstateの変更(Play to Stop)が呼び出されるとErrChangeSessionStateNotPermit(status code: 400)",
-			sessionID:             "sessionID",
-			body:                  `{"state": "STOP"}`,
-			prepareMockPlayerFn:   func(m *mock_spotify.MockPlayer) {},
-			prepareMockPusherFn:   func(m *mock_event.MockPusher) {},
-			prepareMockUserRepoFn: func(m *mock_repository.MockUser) {},
-			prepareMockSessionRepoFn: func(m *mock_repository.MockSession) {
-				m.EXPECT().FindByID("sessionID").Return(&entity.Session{
-					ID:        "sessionID",
-					Name:      "session_name",
-					CreatorID: "creator_id",
-					QueueHead: 0,
-					DeviceID:  "device_id",
-					StateType: entity.Play,
-					QueueTracks: []*entity.QueueTrack{
-						{Index: 0, URI: "spotify:track:5uQ0vKy2973Y9IUCd1wMEF"},
-						{Index: 1, URI: "spotify:track:49BRCNV7E94s7Q2FUhhT3w"},
-					},
-				}, nil)
-			},
-			wantErr:  true,
-			wantCode: http.StatusBadRequest,
-		},
-		{
-			name:      "PLAYで再生するデバイスがオフラインのとき403",
-			sessionID: "sessionID",
-			body:      `{"state": "PLAY"}`,
-			prepareMockPlayerFn: func(m *mock_spotify.MockPlayer) {
-				m.EXPECT().SetRepeatMode(gomock.Any(), false, "device_id").Return(entity.ErrActiveDeviceNotFound)
-			},
-			prepareMockPusherFn:   func(m *mock_event.MockPusher) {},
-			prepareMockUserRepoFn: func(m *mock_repository.MockUser) {},
-			prepareMockSessionRepoFn: func(m *mock_repository.MockSession) {
-				m.EXPECT().FindByID("sessionID").Return(&entity.Session{
-					ID:        "sessionID",
-					Name:      "session_name",
-					CreatorID: "creator_id",
-					QueueHead: 0,
-					DeviceID:  "device_id",
-					StateType: entity.Pause,
-					QueueTracks: []*entity.QueueTrack{
-						{Index: 0, URI: "spotify:track:5uQ0vKy2973Y9IUCd1wMEF"},
-						{Index: 1, URI: "spotify:track:49BRCNV7E94s7Q2FUhhT3w"},
-					},
-				}, nil)
-			},
-			wantErr:  true,
-			wantCode: http.StatusForbidden,
-		},
-		{
-			name:      "PLAYで再生するデバイスがオフラインかつStateがStopなら403",
-			sessionID: "sessionID",
-			body:      `{"state": "PLAY"}`,
-			prepareMockPlayerFn: func(m *mock_spotify.MockPlayer) {
-				m.EXPECT().SetRepeatMode(gomock.Any(), false, "device_id").Return(entity.ErrActiveDeviceNotFound)
-			},
-			prepareMockPusherFn:   func(m *mock_event.MockPusher) {},
-			prepareMockUserRepoFn: func(m *mock_repository.MockUser) {},
-			prepareMockSessionRepoFn: func(m *mock_repository.MockSession) {
-				m.EXPECT().FindByID("sessionID").Return(&entity.Session{
-					ID:        "sessionID",
-					Name:      "session_name",
-					CreatorID: "creator_id",
-					QueueHead: 0,
-					DeviceID:  "device_id",
-					StateType: entity.Stop,
-					QueueTracks: []*entity.QueueTrack{
-						{Index: 0, URI: "spotify:track:5uQ0vKy2973Y9IUCd1wMEF"},
-						{Index: 1, URI: "spotify:track:49BRCNV7E94s7Q2FUhhT3w"},
-					},
-				}, nil)
-			},
-			wantErr:  true,
-			wantCode: http.StatusForbidden,
-		},
-		{
-			name:      "PLAYでキューに一曲も曲が詰まれていないときは400",
-			sessionID: "sessionID",
-			body:      `{"state": "PLAY"}`,
-			prepareMockPlayerFn: func(m *mock_spotify.MockPlayer) {
-				m.EXPECT().SetRepeatMode(gomock.Any(), false, "device_id").Return(nil)
-				m.EXPECT().SetShuffleMode(gomock.Any(), false, "device_id").Return(nil)
-			},
-			prepareMockUserRepoFn: func(m *mock_repository.MockUser) {},
-			prepareMockPusherFn:   func(m *mock_event.MockPusher) {},
-			prepareMockSessionRepoFn: func(m *mock_repository.MockSession) {
-				m.EXPECT().FindByID("sessionID").Return(&entity.Session{
-					ID:          "sessionID",
-					Name:        "session_name",
-					CreatorID:   "creator_id",
-					QueueHead:   0,
-					DeviceID:    "device_id",
-					StateType:   entity.Stop,
-					QueueTracks: nil,
-				}, nil)
-
-			},
-			wantErr:  true,
-			wantCode: http.StatusBadRequest,
-		},
-		{
-			name:      "PLAYでStateTypeがSTOPのときは1曲を再生し次の1曲のみSpotifyのqueueに追加し、再生を始めて202",
-			sessionID: "sessionID",
-			body:      `{"state": "PLAY"}`,
-			prepareMockPlayerFn: func(m *mock_spotify.MockPlayer) {
-				m.EXPECT().SetRepeatMode(gomock.Any(), false, "device_id").Return(nil)
-				m.EXPECT().SetShuffleMode(gomock.Any(), false, "device_id").Return(nil)
-				m.EXPECT().SkipAllTracks(gomock.Any(), "device_id", "spotify:track:5uQ0vKy2973Y9IUCd1wMEF").Return(nil)
-				m.EXPECT().PlayWithTracks(gomock.Any(), "device_id", []string{"spotify:track:5uQ0vKy2973Y9IUCd1wMEF"}).Return(nil)
-				m.EXPECT().AddToQueue(gomock.Any(), "spotify:track:49BRCNV7E94s7Q2FUhhT3w", "device_id").Return(nil)
-				m.EXPECT().AddToQueue(gomock.Any(), "spotify:track:3", "device_id").Return(nil)
-			},
-			prepareMockUserRepoFn: func(m *mock_repository.MockUser) {},
-			prepareMockPusherFn: func(m *mock_event.MockPusher) {
-				m.EXPECT().Push(&event.PushMessage{
-					SessionID: "sessionID",
-					Msg:       entity.EventPlay,
-				})
-			},
-			prepareMockSessionRepoFn: func(m *mock_repository.MockSession) {
-				m.EXPECT().FindByID("sessionID").Return(&entity.Session{
-					ID:        "sessionID",
-					Name:      "session_name",
-					CreatorID: "creator_id",
-					QueueHead: 0,
-					DeviceID:  "device_id",
-					StateType: entity.Stop,
-					QueueTracks: []*entity.QueueTrack{
-						{Index: 0, URI: "spotify:track:5uQ0vKy2973Y9IUCd1wMEF"},
-						{Index: 1, URI: "spotify:track:49BRCNV7E94s7Q2FUhhT3w"},
-						{Index: 2, URI: "spotify:track:3"},
-					},
-				}, nil)
-				m.EXPECT().Update(&entity.Session{
-					ID:        "sessionID",
-					Name:      "session_name",
-					CreatorID: "creator_id",
-					QueueHead: 0,
-					DeviceID:  "device_id",
-					StateType: "PLAY",
-					QueueTracks: []*entity.QueueTrack{
-						{Index: 0, URI: "spotify:track:5uQ0vKy2973Y9IUCd1wMEF"},
-						{Index: 1, URI: "spotify:track:49BRCNV7E94s7Q2FUhhT3w"},
-						{Index: 2, URI: "spotify:track:3"},
-					},
-				}).Return(nil)
-			},
-			wantErr:  false,
-			wantCode: http.StatusAccepted,
-		},
-		{
-			name:      "PLAYでStateTypeがPauseのときはキューはいじらず再生をし始めて202",
-			sessionID: "sessionID",
-			body:      `{"state": "PLAY"}`,
-			prepareMockPlayerFn: func(m *mock_spotify.MockPlayer) {
-				m.EXPECT().SetRepeatMode(gomock.Any(), false, "device_id").Return(nil)
-
-				m.EXPECT().SetShuffleMode(gomock.Any(), false, "device_id").Return(nil)
-				m.EXPECT().Play(gomock.Any(), "device_id").Return(nil)
-			},
-			prepareMockUserRepoFn: func(m *mock_repository.MockUser) {},
-			prepareMockPusherFn: func(m *mock_event.MockPusher) {
-				m.EXPECT().Push(&event.PushMessage{
-					SessionID: "sessionID",
-					Msg:       entity.EventPlay,
-				})
-			},
-			prepareMockSessionRepoFn: func(m *mock_repository.MockSession) {
-				m.EXPECT().FindByID("sessionID").Return(&entity.Session{
-					ID:        "sessionID",
-					Name:      "session_name",
-					CreatorID: "creator_id",
-					QueueHead: 0,
-					DeviceID:  "device_id",
-					StateType: entity.Pause,
-					QueueTracks: []*entity.QueueTrack{
-						{Index: 0, URI: "spotify:track:5uQ0vKy2973Y9IUCd1wMEF"},
-						{Index: 1, URI: "spotify:track:49BRCNV7E94s7Q2FUhhT3w"},
-					},
-				}, nil)
-				m.EXPECT().Update(&entity.Session{
-					ID:        "sessionID",
-					Name:      "session_name",
-					CreatorID: "creator_id",
-					QueueHead: 0,
-					DeviceID:  "device_id",
-					StateType: "PLAY",
-					QueueTracks: []*entity.QueueTrack{
-						{Index: 0, URI: "spotify:track:5uQ0vKy2973Y9IUCd1wMEF"},
-						{Index: 1, URI: "spotify:track:49BRCNV7E94s7Q2FUhhT3w"},
-					},
-				}).Return(nil)
-			},
-			wantErr:  false,
-			wantCode: http.StatusAccepted,
-		},
-		{
-			name:                  "PAUSEで指定されたidのセッションが存在しないとき404",
-			sessionID:             "notFoundSessionID",
-			body:                  `{"state": "PAUSE"}`,
-			prepareMockPlayerFn:   func(m *mock_spotify.MockPlayer) {},
-			prepareMockPusherFn:   func(m *mock_event.MockPusher) {},
-			prepareMockUserRepoFn: func(m *mock_repository.MockUser) {},
-			prepareMockSessionRepoFn: func(m *mock_repository.MockSession) {
-				m.EXPECT().FindByID("notFoundSessionID").Return(nil, entity.ErrSessionNotFound)
-			},
-			wantErr:  true,
-			wantCode: http.StatusNotFound,
-		},
-		{
-			name:      "PAUSEで再生するデバイスがオフラインのときは、既に再生が止まっているはずなので202を返す",
-			sessionID: "sessionID",
-			body:      `{"state": "PAUSE"}`,
-			prepareMockPlayerFn: func(m *mock_spotify.MockPlayer) {
-				m.EXPECT().Pause(gomock.Any(), "device_id").Return(entity.ErrActiveDeviceNotFound)
-			},
-			prepareMockPusherFn: func(m *mock_event.MockPusher) {
-				m.EXPECT().Push(&event.PushMessage{
-					SessionID: "sessionID",
-					Msg:       entity.EventPause,
-				})
-			},
-			prepareMockUserRepoFn: func(m *mock_repository.MockUser) {},
-			prepareMockSessionRepoFn: func(m *mock_repository.MockSession) {
-				m.EXPECT().FindByID("sessionID").Return(&entity.Session{
-					ID:          "sessionID",
-					Name:        "session_name",
-					CreatorID:   "creator_id",
-					DeviceID:    "device_id",
-					QueueHead:   0,
-					StateType:   "PAUSE",
-					QueueTracks: nil,
-				}, nil)
-				m.EXPECT().Update(&entity.Session{
-					ID:          "sessionID",
-					Name:        "session_name",
-					CreatorID:   "creator_id",
-					DeviceID:    "device_id",
-					QueueHead:   0,
-					StateType:   "PAUSE",
-					QueueTracks: nil,
-				}).Return(nil)
-			},
-			wantErr:  false,
-			wantCode: http.StatusAccepted,
-		},
-		{
-			name:                "STOPで正しく再生リクエストが処理されたとき202",
-			sessionID:           "sessionID",
-			body:                `{"state": "STOP"}`,
-			prepareMockPlayerFn: func(m *mock_spotify.MockPlayer) {},
-			prepareMockPusherFn: func(m *mock_event.MockPusher) {
-				m.EXPECT().Push(&event.PushMessage{
-					SessionID: "sessionID",
-					Msg:       entity.EventUnarchive,
-				})
-			},
-			prepareMockUserRepoFn: func(m *mock_repository.MockUser) {},
-			prepareMockSessionRepoFn: func(m *mock_repository.MockSession) {
-				m.EXPECT().FindByID("sessionID").Return(&entity.Session{
-					ID:          "sessionID",
-					Name:        "session_name",
-					CreatorID:   "creator_id",
-					QueueHead:   0,
-					DeviceID:    "device_id",
-					StateType:   "ARCHIVED",
-					QueueTracks: nil,
-				}, nil)
-				m.EXPECT().Update(&entity.Session{
-					ID:          "sessionID",
-					Name:        "session_name",
-					CreatorID:   "creator_id",
-					QueueHead:   0,
-					DeviceID:    "device_id",
-					StateType:   "STOP",
-					QueueTracks: nil,
-				}).Return(nil)
-			},
-			wantErr:  false,
-			wantCode: http.StatusAccepted,
-		},
-		{
-			name:      "ARCHIVEDで正しく再生リクエストが処理されたとき202",
-			sessionID: "sessionID",
-			body:      `{"state": "ARCHIVED"}`,
-			prepareMockPlayerFn: func(m *mock_spotify.MockPlayer) {
-				m.EXPECT().Pause(gomock.Any(), "device_id").Return(nil)
-			},
-			prepareMockPusherFn: func(m *mock_event.MockPusher) {
-				m.EXPECT().Push(&event.PushMessage{
-					SessionID: "sessionID",
-					Msg:       entity.EventArchived,
-				})
-			},
-			prepareMockUserRepoFn: func(m *mock_repository.MockUser) {},
-			prepareMockSessionRepoFn: func(m *mock_repository.MockSession) {
-				m.EXPECT().FindByID("sessionID").Return(&entity.Session{
-					ID:          "sessionID",
-					Name:        "session_name",
-					CreatorID:   "creator_id",
-					QueueHead:   0,
-					DeviceID:    "device_id",
-					StateType:   "PLAY",
-					QueueTracks: nil,
-				}, nil)
-				m.EXPECT().Update(&entity.Session{
-					ID:          "sessionID",
-					Name:        "session_name",
-					CreatorID:   "creator_id",
-					QueueHead:   0,
-					DeviceID:    "device_id",
-					StateType:   "ARCHIVED",
-					QueueTracks: nil,
-				}).Return(nil)
-			},
-			wantErr:  false,
-			wantCode: http.StatusAccepted,
-		},
-		{
-			name:      "PAUSEで正しく再生リクエストが処理されたとき202",
-			sessionID: "sessionID",
-			body:      `{"state": "PAUSE"}`,
-			prepareMockPlayerFn: func(m *mock_spotify.MockPlayer) {
-				m.EXPECT().Pause(gomock.Any(), "device_id").Return(nil)
-			},
-			prepareMockPusherFn: func(m *mock_event.MockPusher) {
-				m.EXPECT().Push(&event.PushMessage{
-					SessionID: "sessionID",
-					Msg:       entity.EventPause,
-				})
-			},
-			prepareMockUserRepoFn: func(m *mock_repository.MockUser) {},
-			prepareMockSessionRepoFn: func(m *mock_repository.MockSession) {
-				m.EXPECT().FindByID("sessionID").Return(&entity.Session{
-					ID:          "sessionID",
-					Name:        "session_name",
-					CreatorID:   "creator_id",
-					QueueHead:   0,
-					DeviceID:    "device_id",
-					StateType:   "PAUSE",
-					QueueTracks: nil,
-				}, nil)
-				m.EXPECT().Update(&entity.Session{
-					ID:          "sessionID",
-					Name:        "session_name",
-					CreatorID:   "creator_id",
-					QueueHead:   0,
-					DeviceID:    "device_id",
-					StateType:   "PAUSE",
-					QueueTracks: nil,
-				}).Return(nil)
-			},
-			wantErr:  false,
-			wantCode: http.StatusAccepted,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// httptestの準備
-			e := echo.New()
-			req := httptest.NewRequest(http.MethodPut, "/", strings.NewReader(tt.body))
-			req.Header.Set(echo.HeaderContentType, echo.MIMEApplicationJSON)
-			rec := httptest.NewRecorder()
-			c := e.NewContext(req, rec)
-			c.SetPath("/sessions/:id/state")
-			c.SetParamNames("id")
-			c.SetParamValues(tt.sessionID)
-			c = setToContext(c, tt.userID, nil)
-
-			// モックの準備
-			ctrl := gomock.NewController(t)
-			defer ctrl.Finish()
-			mockPlayer := mock_spotify.NewMockPlayer(ctrl)
-			tt.prepareMockPlayerFn(mockPlayer)
-			mockPusher := mock_event.NewMockPusher(ctrl)
-			tt.prepareMockPusherFn(mockPusher)
-			mockUserRepo := mock_repository.NewMockUser(ctrl)
-			tt.prepareMockUserRepoFn(mockUserRepo)
-			mockSessionRepo := mock_repository.NewMockSession(ctrl)
-			tt.prepareMockSessionRepoFn(mockSessionRepo)
-			uc := usecase.NewSessionUseCase(mockSessionRepo, mockUserRepo, mockPlayer, nil, nil, mockPusher)
-			h := &SessionHandler{
-				uc: uc,
-			}
-			err := h.State(c)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("State() error = %v, wantErr %v", err, tt.wantErr)
-			}
-
-			// ステータスコードのチェック
-			if er, ok := err.(*echo.HTTPError); ok && er.Code != tt.wantCode {
-				t.Errorf("State() code = %d, want = %d", rec.Code, tt.wantCode)
-			}
-		})
-	}
-}
-
 func TestSessionHandler_SetDevice(t *testing.T) {
 	t.Parallel()
